# apps/api/src/apps/roles/dto/create-role.dto.ts
import { ApiProperty } from "@nestjs/swagger";
import { Transform } from "class-transformer";
import { IsNotEmpty, IsString } from "class-validator";
import type { CreateRoleRequest } from "@thanks-claude/shared-types";

export class CreateRoleDto implements CreateRoleRequest {
  @ApiProperty()
  @Transform(({ value }) => (typeof value === "string" ? value.trim() : value))
  @IsString()
  @IsNotEmpty()
  name!: string;
}